+++ core/gui_mbox.c
@@ -34,7 +34,7 @@ static struct {
         { MBOX_BTN_CANCEL,  LANG_MBOX_BTN_CANCEL}
 };
 #define BUTTON_SIZE     6
-#define BUTTONSNUM      (sizeof(buttons)/sizeof(buttons[0]))
+#define BUTTONSNUM      (int)(sizeof(buttons)/sizeof(buttons[0]))
 #define MAX_BUTTONS     3
 static int      mbox_buttons[MAX_BUTTONS], mbox_buttons_num, mbox_button_active;
 static coord    mbox_buttons_x, mbox_buttons_y;
